server/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
charset-normalizer==2.0.9
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
idna==3.3
mysqlclient==2.1.0
Pillow==8.4.0
pycodestyle==2.8.0
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.7
